ThreadedWindow imports MainWindow, cleaned up comments
